--- Engineering/EngineeringTeam.cpp
@@ -1,5 +1,0 @@
-#include "EngineeringTeam.h"
-
-void EngineeringTeam::changed() {
-	_sendSpecs->relayDimensions(this);
-}
