+++ Mediator/EngineeringTeam.cpp
@@ -1,16 +1,5 @@
 #include "EngineeringTeam.h"
 
 void EngineeringTeam::changed() {
-	// TODO - implement EngineeringTeam::changed
-	throw "Not yet implemented";
-}
-
-sendSpecs EngineeringTeam::getSpecs() {
-	// TODO - implement EngineeringTeam::getSpecs
-	throw "Not yet implemented";
-}
-
-void EngineeringTeam::set(sendSpecs s) {
-	// TODO - implement EngineeringTeam::set
-	throw "Not yet implemented";
+	_sendSpecs->relayDimensions(this);
 }
